pyproject.toml
[project]
name = "short-language-detection"
version = "0.1.2"
description = "A simple language detection library for short texts."
license = "MIT"
readme = "README.md"
authors = [
    { name = "jourdelune", email = "jourdelune863@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "emoji>=2.14.1",
    "fasttext-wheel>=0.9.2",
    "lingua-language-detector>=2.0.2",
    "numpy==1.26.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
shortlanguagedetection = { git = "https://github.com/Interaction-Bot/LanguageDetection.git" }
